extract metadata to file
